backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.18
pydantic>=2.5.0
python-dotenv>=1.0.0
openai>=1.35.0
cohere>=4.37
pinecone>=5.0.0
pymongo>=4.6.3
tiktoken>=0.5.2
numpy>=1.24.3
python-jose[cryptography]>=3.3.0
httpx>=0.25.2
setuptools>=78.1.1
pypdf2
python-docx